Refactored code for data to be pulled in main component
